winbindd: support additional groups in GETGROUPS
Allow the winbindd WINBINDD_GETGROUPS API to return an
additional set of UNIX groups, in accordance with the new
"winbind additional group sids" parameter.

The idea is to add the SIDs to the list of SIDs obtained
from active directory, and then with a suitable group mapping
map those SIDs to UNIX ids.

Signed-off-by: Uri Simchoni <urisimchoni@gmail.com>