winbindd sids2xids: separate resolving list from mapping list
Manage two lists instead of one - a list of SIDs to be mapped,
and a (conceptual) subset that needs to be resolved before mapping.

In this patch we just introduce the lists and they have the same
content, but in later patches we can remove some SIDs from the
resolve list.

Fixes APP-13427